Move DHCP from network into its own service
