+++ main.go
@@ -9,6 +9,7 @@ import (
 	"github.com/rancher/docker-from-scratch"
 	"github.com/rancher/os/cmd/cloudinit"
 	"github.com/rancher/os/cmd/control"
+	"github.com/rancher/os/cmd/dhcp"
 	"github.com/rancher/os/cmd/network"
 	"github.com/rancher/os/cmd/power"
 	"github.com/rancher/os/cmd/respawn"
@@ -16,7 +17,6 @@ import (
 	"github.com/rancher/os/cmd/systemdocker"
 	"github.com/rancher/os/cmd/userdocker"
 	"github.com/rancher/os/cmd/wait"
-	"github.com/rancher/os/cmd/waitfornetwork"
 	"github.com/rancher/os/config"
 	osInit "github.com/rancher/os/init"
 )
@@ -53,7 +53,7 @@ func main() {
 	registerCmd("/usr/bin/ros", control.Main)
 	registerCmd("/usr/bin/cloud-init", cloudinit.Main)
 	registerCmd("/usr/sbin/netconf", network.Main)
-	registerCmd("/usr/sbin/wait-for-network", waitfornetwork.Main)
+	registerCmd("/usr/sbin/dhcp", dhcp.Main)
 	registerCmd("/usr/sbin/wait-for-docker", wait.Main)
 
 	if !reexec.Init() {
